backend/requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.5
colorama==0.4.6
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
Django==4.2.1
djangorestframework==3.14.0
exceptiongroup==1.1.1
executing==1.2.0
Faker==18.9.0
fastapi==0.100.0
ffmpy==0.3.0
filelock==3.12.2
fonttools==4.41.0
frozenlist==1.4.0
fsspec==2023.6.0
gradio==3.36.1
gradio_client==0.2.9
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
ipykernel==6.23.3
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
linkify-it-py==2.0.2
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mysqlclient==2.1.1
nest-asyncio==1.5.6
numpy==1.24.3
openai==0.27.8
orjson==3.9.2
outcome==1.2.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.8.0
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
pycparser==2.21
pydantic==2.0.2
pydantic_core==2.1.2
pydub==0.25.1
Pygments==2.15.1
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.6
python-nmap==0.7.1
pytz==2023.3
pywin32==306
PyYAML==6.0
pyzmq==25.1.0
referencing==0.29.1
requests==2.31.0
rpds-py==0.8.10
scipy==1.10.1
selenium==4.10.0
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
speedtest==0.0.1
speedtest-cli==2.1.3
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.3
typing_extensions==4.7.1
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==2.0.3
uvicorn==0.22.0
wcwidth==0.2.6
websockets==11.0.3
wsproto==1.2.0
yarl==1.9.2
